Stop inserting spurious newline with --separate-relative
Until now, if the module contained only relative imports and
`--separate-relative` was passed, then a spurious newline was added at the
beginnning of the file. With this commit the newline is added only when
there are additional imports.

This is the second attempt to fix this. The previous attempt was e9b44a4
which caused a regression described in #136.

Fixes #134.